separated retrieval metric test files and removed useless loops
