+++ tests/functional/test_retrieval.py
@@ -6,8 +6,8 @@ import torch
 from pytorch_lightning import seed_everything
 from sklearn.metrics import average_precision_score as sk_average_precision
 
-from tests.retrieval.utils import _assert_error
-from tests.retrieval.utils import _reciprocal_rank as reciprocal_rank
+from tests.retrieval.test_mrr import _reciprocal_rank as reciprocal_rank
+from tests.retrieval.test_utils import _assert_error
 from torchmetrics.functional.retrieval.average_precision import retrieval_average_precision
 from torchmetrics.functional.retrieval.reciprocal_rank import retrieval_reciprocal_rank
 
@@ -23,7 +23,7 @@ def test_metrics_output_values(sklearn_metric, torch_metric, size):
     seed_everything(0)
 
     # test results are computed correctly wrt std implementation
-    for i in range(10):
+    for i in range(6):
         preds = np.random.randn(size)
         target = np.random.randn(size) > 0
 
@@ -78,8 +78,7 @@ def test_input_shapes(torch_metric) -> None:
     _assert_error(torch_metric, ValueError, preds, target)
 
     # test checks when shapes are different
-    for _ in range(10):
-        elements_1, elements_2 = np.random.choice(np.arange(1, 20), size=2, replace=False)  # ensure sizes are different
-        preds = torch.tensor([0] * elements_1, device=device, dtype=torch.float)
-        target = torch.tensor([0] * elements_2, device=device, dtype=torch.int64)
-        _assert_error(torch_metric, ValueError, preds, target)
+    elements_1, elements_2 = np.random.choice(np.arange(1, 20), size=2, replace=False)  # ensure sizes are different
+    preds = torch.tensor([0] * elements_1, device=device, dtype=torch.float)
+    target = torch.tensor([0] * elements_2, device=device, dtype=torch.int64)
+    _assert_error(torch_metric, ValueError, preds, target)
